a simple authentication system initialized

+++ app.py
@@ -18,10 +18,6 @@ def validate_user(request_body):
     if "role" not in request_body:
         request_body['role'] = "ANNOTATOR"
 
-    # if "token" does not exist in request_body, set it to ""
-    if "token" not in request_body:
-        request_body['token'] = ""
-
     if request_body['role'] not in ["ADMIN", "ANNOTATOR", "REVIEWER"]:
         return 2
 
@@ -29,8 +25,25 @@ def validate_user(request_body):
     request_body['number_of_completed_projects'] = 0
 
 
+@app.route('/login', methods=['POST'])
+def login():
+    # Check if username and password are in request body
+    if not all([field in request.json for field in ['username', 'password']]):
+        return jsonify({"status": "username or password fields are missing"})
+    # Check if username and password are correct
+    user = db.Users.find_one(
+        {'username': request.json['username'], 'password': request.json['password']})
+    if user is None:
+        return jsonify({"status": "username or password is incorrect"})
+    else:
+        return jsonify({"role": str(user['role'])})
+
+
 @app.route('/Users/', methods=['POST'])
 def create_user():
+    if request.headers.get('role') != "ADMIN":
+        return jsonify({"status": "Only ADMIN can create users"})
+
     error_code = validate_user(request.json)
     if (1 == error_code):
         return jsonify({"status": "username or password fields are missing"})
@@ -44,6 +57,9 @@ def create_user():
 
 @app.route('/Users/', methods=['GET'])
 def show_users():
+    if request.headers.get('role') != "ADMIN":
+        return jsonify({"status": "Only ADMIN can see users"})
+
     users = []
     for user in db.Users.find():
         user['_id'] = str(user['_id'])
@@ -53,6 +69,9 @@ def show_users():
 
 @app.route('/Users/<user_id>', methods=['PUT'])
 def update_user(user_id):
+    if request.headers.get('role') != "ADMIN":
+        return jsonify({"status": "Only ADMIN can update users"})
+
     try:
         error_code = validate_user(request.json)
         if (1 == error_code):
@@ -78,6 +97,9 @@ def update_user(user_id):
 
 @app.route('/Users/<user_id>', methods=['DELETE'])
 def delete_user(user_id):
+    if request.headers.get('role') != "ADMIN":
+        return jsonify({"status": "Only ADMIN can delete users"})
+
     try:
         response = db.Users.delete_one({'_id': ObjectId(user_id)})
         if response.deleted_count == 0:
